feat(domain): implement class car
Added methods:
- validate
- move forward
