+++ src/main/java/racingcar/domain/Car.java
@@ -1,0 +1,32 @@
+package racingcar.domain;
+
+public class Car {
+    private static final int NAME_LENGTH_MIN = 1;
+    private static final int NAME_LENGTH_MAX = 5;
+    private final String name;
+    private int position = 0;
+
+    public Car(String name) {
+        validate(name);
+        this.name = name;
+    }
+
+    private void validate(String name) {
+        if (name.length() < NAME_LENGTH_MIN || name.length() > NAME_LENGTH_MAX) {
+            throw new IllegalArgumentException(String.format(ErrorMessage.CAR_NAME_LENGTH.message(),
+                    NAME_LENGTH_MIN, NAME_LENGTH_MAX));
+        }
+    }
+
+    public void moveForward() {
+        position++;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPosition() {
+        return position;
+    }
+}
